Renamed a few packages to internal in the minecraft binding. See openhab/static-code-analysis#100
