+++ addons/binding/org.openhab.binding.minecraft/src/main/java/org/openhab/binding/minecraft/handler/MinecraftServerHandler.java
@@ -20,13 +20,13 @@ import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.State;
 import org.openhab.binding.minecraft.MinecraftBindingConstants;
-import org.openhab.binding.minecraft.config.ServerConfig;
-import org.openhab.binding.minecraft.handler.server.ServerConnection;
-import org.openhab.binding.minecraft.message.OHMessage;
-import org.openhab.binding.minecraft.message.data.PlayerData;
-import org.openhab.binding.minecraft.message.data.ServerData;
-import org.openhab.binding.minecraft.message.data.SignData;
-import org.openhab.binding.minecraft.util.RetryWithDelay;
+import org.openhab.binding.minecraft.internal.config.ServerConfig;
+import org.openhab.binding.minecraft.internal.message.OHMessage;
+import org.openhab.binding.minecraft.internal.message.data.PlayerData;
+import org.openhab.binding.minecraft.internal.message.data.ServerData;
+import org.openhab.binding.minecraft.internal.message.data.SignData;
+import org.openhab.binding.minecraft.internal.server.ServerConnection;
+import org.openhab.binding.minecraft.internal.util.RetryWithDelay;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
