+++ addons/binding/org.openhab.binding.minecraft/src/main/java/org/openhab/binding/minecraft/internal/message/data/PlayerData.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.minecraft.message.data;
+package org.openhab.binding.minecraft.internal.message.data;
 
 /**
  * Object representing Minecraft player.
