+++ addons/binding/org.openhab.binding.minecraft/src/main/java/org/openhab/binding/minecraft/internal/message/data/commands/PlayerCommandData.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.minecraft.message.data.commands;
+package org.openhab.binding.minecraft.internal.message.data.commands;
 
 /**
  * Object representing Minecraft server.
