+++ addons/binding/org.openhab.binding.minecraft/src/main/java/org/openhab/binding/minecraft/internal/util/RetryWithDelay.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.minecraft.util;
+package org.openhab.binding.minecraft.internal.util;
 
 import java.util.concurrent.TimeUnit;
